Update unifi_protect_backup_core.py
Fix typo in connection attempts.  The application only attempts to connect once instead of 10 times.  Incidentally the newest version of Unifi Protect is not fully compatible with pyunifiprotect causing an exception to be thrown on the first attempt, but not the second.  This also will allow the application to continue working with the newest version of pyunifiprotect.